Added Activity Timer - Mouse activity detected
The mouse and keyboard activity are now detected on a fairly tight loop
and tested to show the player control panel on activity.  Activity is
detected system wide. (We can limit this in the future if needs be)
GetLastInput is now available to give a TimeSpan of idle time.  Could be
wrapped in a field.